refactor: remove ChainFunding trait The ChainFunding trait in the shared `chain` crate was a test-only abstraction with a single caller: MemWallet::funded_wallet_from_rpc,used by the rpc bmp_service integration test. It had one real  impl (BitcoinCoreChainApi, local to that test) and one dead stub (Testchain, both methods bail!). Delete the trait, drop funded_wallet_from_rpc from the wallet crate, and inline the funding bootstrap as a local helper in the test that calls bitcoind RPC directly.
